Update index.rst
In general it could be helpful to have a rundown of which parts of the code will be run only if the model is uncoupled and which parts run only if the model IS coupled. Plus a quick list of which namelist parameters HAVE to be set in a certain way and which ones become obsolete.
e.g. Salinity and Temperature restoring have to be turned off.