Update getting_started.rst
Hi, I figured out that when performing a full second cycle starting at the very first year of the forcing (e.g. 1958 in JRA55) the 'yearnew' parameter in 'namelist.config' needs to be changed to the previous year (in this case 1957), otherwise the model will perform an initialization run instead of a restart. Maybe this should be added to the docs.